Add files via upload
# Introduction
In this file, MVPA is applied on Z-stat files from first level analysis I have done on naive space, ART applied images, with each stimuli giving one Z-stat image.

MPVA is applied for all voxels of the brain. For each participant, data is split into the number of runs, with data from one run is used for test, and the rest is used for training. 

## Future works
Optimizing the results based on various types of classifiers, working on classifier performance are works to be done.