MVPA for spotlight search
Around each voxel, a sphere with a predefined radius is fit as spotlight. The voxels within this sphere is the input to the classifier. As the classification is required for each brain voxel, parallel processing significantly impacts processing time, which has been implemented in this version.

Tasks to do: save the nii output file, run for all participants, compare results, optimize ML parameters, sphere radius, etc.